Refactor to other files
