+++ packages/common/src/api/tan-query/comments/useGetCommentById.ts
@@ -1,0 +1,43 @@
+import { useEffect } from 'react'
+
+import { useQuery, useQueryClient } from '@tanstack/react-query'
+import { useDispatch } from 'react-redux'
+
+import { useAudiusQueryContext } from '~/audius-query'
+import { Feature, ID } from '~/models'
+import { toast } from '~/store/ui/toast/slice'
+
+import { CommentOrReply, messages } from './types'
+import { getCommentQueryKey } from './utils'
+
+export const useGetCommentById = (commentId: ID) => {
+  const { reportToSentry } = useAudiusQueryContext()
+  const queryClient = useQueryClient()
+  const dispatch = useDispatch()
+
+  const queryRes = useQuery({
+    queryKey: getCommentQueryKey(commentId),
+    enabled: !!commentId,
+    queryFn: async (): Promise<CommentOrReply | {}> => {
+      // TODO: there's no backend implementation of this fetch at the moment;
+      // but we also never expect to call the backend here; we always prepopulate the data from the fetch by tracks method
+      return queryClient.getQueryData(getCommentQueryKey(commentId)) ?? {}
+    },
+    staleTime: Infinity
+  })
+
+  const { error } = queryRes
+
+  useEffect(() => {
+    if (error) {
+      reportToSentry({
+        error,
+        name: 'Comments',
+        feature: Feature.Comments
+      })
+      dispatch(toast({ content: messages.loadError('comments') }))
+    }
+  }, [error, dispatch, reportToSentry])
+
+  return queryRes
+}
